moved common class and removed utils.py
